editor

+++ src/free_game.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   free_game.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: apavlov <apavlov@student.unit.ua>          +#+  +:+       +#+        */
+/*   By: myuliia <myuliia@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/13 22:19:24 by apavlov           #+#    #+#             */
-/*   Updated: 2019/08/13 22:19:25 by apavlov          ###   ########.fr       */
+/*   Updated: 2019/08/22 17:24:26 by myuliia          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void		free_file_data(t_map *map)
 	int		i;
 
 	i = -1;
-	while (++i < map->num_sect)
+	while (++i < (int)map->num_sect)
 	{
 		free(map->sectors[i].vert);
 		free(map->sectors[i].neighbors);
